+++ src/OpenRasta.Tests.Unit/TypeSystem/Members_Specification.cs
@@ -73,7 +73,7 @@ namespace Accessors_Specification
         [Test, Ignore("no idea")]
         public void interfaces_cannot_be_initialized()
         {
-            Executing(GivenTypeFor<ICollection>)
+            Executing((Action)GivenTypeFor<ICollection>)
                 .ShouldThrow<NotSupportedException>();
         }