+++ src/OpenRasta/OperationModel/IOperationExecutor.cs
@@ -1,10 +1,11 @@
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using OpenRasta.Web;
 
 namespace OpenRasta.OperationModel
 {
     public interface IOperationExecutor
     {
-        OperationResult Execute(IEnumerable<IOperation> operations);
+        Task<OperationResult> Execute(IEnumerable<IOperation> operations);
     }
 }