+++ src/OpenRasta/Pipeline/Contributors/OperationInvokerContributor.cs
@@ -1,4 +1,6 @@
-﻿using OpenRasta.DI;
+﻿using System;
+using System.Threading.Tasks;
+using OpenRasta.DI;
 using OpenRasta.OperationModel;
 using OpenRasta.OperationModel.Interceptors;
 using OpenRasta.Web;
@@ -6,35 +8,34 @@ using OpenRasta.Pipeline;
 
 namespace OpenRasta.Pipeline.Contributors
 {
-    public class OperationInvokerContributor : KnownStages.IOperationExecution
+  public class OperationInvokerContributor : KnownStages.IOperationExecution
+  {
+    readonly IDependencyResolver _resolver;
+
+    public OperationInvokerContributor(IDependencyResolver resolver)
     {
-        readonly IDependencyResolver _resolver;
-
-        public OperationInvokerContributor(IDependencyResolver resolver)
-        {
-            _resolver = resolver;
-        }
-
-        public void Initialize(IPipeline pipelineRunner)
-        {
-            pipelineRunner.Notify(ExecuteOperations).After<KnownStages.IRequestDecoding>();
-        }
-
-        PipelineContinuation ExecuteOperations(ICommunicationContext context)
-        {
-            var executor = _resolver.Resolve<IOperationExecutor>();
-            try
-            {
-
-                context.OperationResult = executor.Execute(context.PipelineData.Operations);
-            }
-            catch (InterceptorException)
-            {
-                if (context.OperationResult != null)
-                    return PipelineContinuation.RenderNow;
-                throw;
-            }
-            return PipelineContinuation.Continue;
-        }
+      _resolver = resolver;
     }
+
+    public void Initialize(IPipeline pipelineRunner)
+    {
+      pipelineRunner.Use(ExecuteOperations).After<KnownStages.IRequestDecoding>();
+    }
+
+    async Task<PipelineContinuation> ExecuteOperations(ICommunicationContext context)
+    {
+      var executor = _resolver.Resolve<IOperationExecutor>();
+      try
+      {
+        context.OperationResult = await executor.Execute(context.PipelineData.Operations);
+      }
+      catch (InterceptorException)
+      {
+        if (context.OperationResult != null)
+          return PipelineContinuation.RenderNow;
+        throw;
+      }
+      return PipelineContinuation.Continue;
+    }
+  }
 }