+++ src/OpenRasta/Pipeline/IPipelineBuilder.cs
@@ -6,7 +6,6 @@ namespace OpenRasta.Pipeline
 {
   public interface IPipelineBuilder
   {
-    IPipelineExecutionOrder Notify(Func<ICommunicationContext, Task<PipelineContinuation>> action);
-    IPipelineExecutionOrder Notify(Func<ICommunicationContext, Task> action);
+    IPipelineExecutionOrder Use(Func<ICommunicationContext, Task<PipelineContinuation>> action);
   }
 }